+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to fastagent are documented in this file.
 
-## Unreleased
+## 0.8.0 — May 9, 2026
 
 ### Bug fixes